# docs/openapi.yml
openapi: 3.0.3
info:
  title: Monorepo Service API
  version: 1.0.0
  description: >-
    HTTP API implemented by apps/service. Protected routes require an access token
    and active persisted session. Routes without explicit role metadata default
    to SuperAdministrator or Administrator. The service also exposes a separate
    WebSocket echo server on port 8080; WebSocket events are not described here.
...
security:
  - bearerAuth: []
paths:
...
  /rent-provider:
    post:
      tags: [Rent providers]
      operationId: createRentProvider
      summary: Create a rent provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentProviderWrite'
      responses:
        '201':
          description: Created rent provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Rent providers]
      operationId: listRentProviders
      summary: List rent providers with pagination
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Paginated rent providers belonging to the authenticated account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentProviderPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /rent-provider/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Rent providers]
      operationId: getRentProvider
      summary: Get an account-owned rent provider by id
      responses:
        '200':
          description: Rent provider or null when no row belongs to the authenticated account.
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/RentProvider'
                nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Rent providers]
      operationId: updateRentProvider
      summary: Update an account-owned rent provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentProviderWrite'
      responses:
        '200':
          description: Updated rent provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Rent providers]
      operationId: deleteRentProvider
      summary: Delete an account-owned rent provider
      responses:
        '200':
          description: Deleted rent provider instance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /invoices:
    post:
      tags: [Invoices]
      operationId: createInvoice
      summary: Create a draft invoice with expense snapshots
      x-required-roles: [SuperAdministrator, Administrator, LocationOperator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreateRequest'
      responses:
        '201':
          description: Created draft invoice and expense snapshots.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceWithExpenses'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Invoices]
      operationId: listInvoices
      summary: List invoices ordered by creation time
      x-required-roles: [SuperAdministrator, Administrator, LocationOperator]
      responses:
        '200':
          description: Invoices.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /invoices/{id}/status:
    patch:
      tags: [Invoices]
      operationId: updateInvoiceStatus
      summary: Mark a draft invoice as done
      description: Only the DRAFT to DONE transition is allowed; completed invoices cannot be reopened.
      x-required-roles: [SuperAdministrator, Administrator]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [DRAFT, DONE]
      responses:
        '200':
          description: Updated invoice.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /invoices/schedules:
    get:
      tags: [Invoices]
      operationId: listInvoiceSchedules
      summary: List locations with schedules and draft invoices
      x-required-roles: [SuperAdministrator, Administrator, LocationOperator]
      responses:
        '200':
          description: One schedule summary per location.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationScheduleSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /invoices/schedules/{locationCode}:
    put:
      tags: [Invoices]
      operationId: saveInvoiceSchedule
      summary: Create or update a location invoice schedule
      x-required-roles: [SuperAdministrator, Administrator, LocationOperator]
      parameters:
        - name: locationCode
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceScheduleWrite'
      responses:
        '200':
          description: Saved invoice schedule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /invoices/schedules/{locationCode}/notify:
    post:
      tags: [Invoices]
      operationId: notifyInvoiceSchedule
      summary: Send the scheduled invoice notification to its rent provider
      x-required-roles: [SuperAdministrator, Administrator, LocationOperator]
      parameters:
        - name: locationCode
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Notification delivery result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /invoices/get-summerize-data/{locationId}:
    get:
      tags: [Invoices]
      operationId: getInvoiceSummaryByLocation
      summary: Get location, tenants, and rent provider summary
      x-required-roles: [SuperAdministrator, Administrator, LocationOperator]
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Location summary data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationInvoiceSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /invoices/{tenantId}:
    get:
      tags: [Invoices]
      operationId: getTenantInvoiceSummary
      summary: Get tenant and assigned location summary
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tenant and, when assigned, its location and rent provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantInvoiceSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by POST /auth/login; active persisted session required.
    cronBearer:
      type: apiKey
      in: header
      name: Authorization
      description: Exact value is Bearer followed by the configured CRON_SECRET.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
  responses:
    BadRequest:
      description: Invalid request or business-rule violation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, or unauthorized credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        msg:
          type: string
...
    PageMetadata:
      type: object
      required: [total, page, size, data]
      properties:
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
    Tenant:
      type: object
      additionalProperties: true
      properties:
        tenantCode:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        tenantName:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: integer
        phoneNumber:
          type: string
        email:
          type: string
          format: email
        contactAddress:
          type: string
        description:
          type: string
        contractUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LocationExpense:
      type: object
      properties:
        expenseCode:
          type: integer
        expenseName:
          type: string
        initialUnit:
          type: number
        currentUnit:
          type: number
        unitName:
          type: string
        type:
          type: string
        price:
          type: number
        inUsed:
          type: boolean
    Location:
      type: object
      additionalProperties: true
      properties:
        locationCode:
          oneOf:
            - type: integer
            - type: string
        locationName:
          type: string
        locationAddress:
          type: string
        roomSize:
          type: number
        description:
          type: string
        owner:
          oneOf:
            - type: integer
            - type: string
        image:
          type: string
          format: uri
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/LocationExpense'
...
    RentProvider:
      type: object
      additionalProperties: true
      properties:
        providerCode:
          type: integer
        providerName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        phoneNumber:
          type: string
        email:
          type: string
          format: email
        contactAddress:
          type: string
        gender:
          type: integer
        description:
          type: string
        accountId:
          type: integer
          readOnly: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RentProviderWrite:
      allOf:
        - $ref: '#/components/schemas/RentProvider'
    RentProviderPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/RentProvider'
    InvoiceStatus:
      type: string
      enum: [DRAFT, DONE]
    InvoiceExpenseSnapshot:
      type: object
      properties:
        expenseCode:
          type: integer
        expenseName:
          type: string
        type:
          type: string
        unitName:
          type: string
        initialUnit:
          type: number
        currentUnit:
          type: number
        price:
          type: number
        unitPrice:
          type: number
        amount:
          type: number
    InvoiceCreateRequest:
      type: object
      required: [locationCode, expenses]
      properties:
        locationCode:
          type: integer
        expenses:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InvoiceExpenseSnapshot'
    Invoice:
      type: object
      additionalProperties: true
      properties:
        invoiceCode:
          type: integer
        locationCode:
          type: integer
        totalAmount:
          type: number
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InvoiceWithExpenses:
      allOf:
        - $ref: '#/components/schemas/Invoice'
        - type: object
          properties:
            expenses:
              type: array
              items:
                $ref: '#/components/schemas/InvoiceExpenseSnapshot'
    InvoiceSchedule:
      type: object
      properties:
        locationCode:
          type: integer
        invoiceCode:
          type: integer
          nullable: true
        dueDay:
          type: integer
        enabled:
          type: boolean
        lastNotifiedAt:
          type: string
          format: date-time
          nullable: true
        lastStatus:
          type: string
          nullable: true
        lastError:
          type: string
          nullable: true
    InvoiceScheduleWrite:
      type: object
      required: [dueDay]
      properties:
        invoiceCode:
          type: integer
          nullable: true
        dueDay:
          type: integer
          minimum: 1
          maximum: 28
        enabled:
          type: boolean
          default: true
    LocationScheduleSummary:
      type: object
      properties:
        location:
          $ref: '#/components/schemas/Location'
        schedule:
          type: object
          allOf:
            - $ref: '#/components/schemas/InvoiceSchedule'
          nullable: true
        invoices:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
    LocationInvoiceSummary:
      type: object
      properties:
        location:
          $ref: '#/components/schemas/Location'
        tenants:
          type: array
          items:
            $ref: '#/components/schemas/Tenant'
        owner:
          type: object
          allOf:
            - $ref: '#/components/schemas/RentProvider'
          nullable: true
    TenantInvoiceSummary:
      type: object
      properties:
        tenant:
          type: object
          allOf:
            - $ref: '#/components/schemas/Tenant'
          nullable: true
        location:
          type: object
          allOf:
            - $ref: '#/components/schemas/Location'
          nullable: true
        owner:
          type: object
          allOf:
            - $ref: '#/components/schemas/RentProvider'
          nullable: true
    NotificationResult:
      type: object
      properties:
        locationCode:
          type: integer
        invoiceCode:
          type: integer
          nullable: true
        recipient:
          type: string
          example: rent-provider
        tenantCount:
          type: integer
        status:
          type: string
          enum: [SENT, FAILED]
        message:
          type: string
...